improved functionalities

+++ app/src/main/java/com/vrsabu/markme/data/remote/models/AttendanceResponse.kt
@@ -60,6 +60,6 @@ data class OverallStatsResponse(
     val totalSessions: Int?,
     val totalAttendanceRecords: Int?,
     val totalPresent: Int?,
-    val overallAttendancePercentage: Int?
+    val overallAttendancePercentage: Float?
 )
